KS-1-4 changing the data type and refactoring code

+++ src/main/kotlin/lesson_1/lesson1_task4.kt
@@ -3,16 +3,15 @@ package org.example.lesson_1
 fun main() {
     val lenght: Long = 40_868_600_000
     val age: Byte = 27
-    val partOfDay: Double = 0.075
-    val seconds: Int = 6480
-    val partOfYear: Double = 2.0547945205479453E-4
+    val partOfDay: Float = 0.075F
+    val seconds: Short = 6480
+    val partOfYear: Float = 2.0547945205479453E-4F
     val apogee: Int = 327000
+
     println("Расстояние: $lenght")
     println("Возраст: $age")
     println("Часть дня: $partOfDay")
     println("Секунды: $seconds")
     println("Часть года: $partOfYear")
     println("Апогей орбиты: $apogee")
-
-
 }
